requirements.txt
absl-py==0.10.0
alabaster==0.7.12
argon2-cffi==20.1.0
asgiref==3.2.10
astor==0.8.1
atomicwrites==1.4.0
attrs==20.1.0
autobahn==19.5.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bleach==1.5.0
certifi==2021.5.30
cffi==1.14.2
channels==2.3.1
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
daphne==2.4.1
dask==2021.3.0
decorator==5.0.9
defusedxml==0.6.0
Django==2.2.16
djangorestframework==3.9.4
docutils==0.16
easydict==1.9
entrypoints==0.3
enum34==1.1.10
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.17
grpcio==1.31.0
h5py==2.10.0
html5lib==0.9999999
hyperlink==20.0.1
idna==2.10
imagecodecs==2020.5.30
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
imgviz==1.2.2
importlib-metadata==1.7.0
incremental==17.5.0
iniconfig==1.1.1
innvestigate==1.0.8
ipykernel==5.3.4
ipyparallel==6.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonfield2==3.0.3
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.0.2
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme==4.5.6
llvmlite==0.36.0
lz4==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mil-pytorch==0.0.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.0
mkl-service==2.3.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
nest==0.1.1
networkx==2.5
nose==1.3.7
notebook==6.1.3
numba==0.53.1
numpy==1.19.5
olefile==0.46
opencv-python==4.4.0.42
openslide-python==1.1.2
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pycparser==2.20
pydensecrf==1.0rc2
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyrsistent==0.16.0
pytest==6.1.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
roman==3.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.3.2
selective-search==1.0.0
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
style==1.1.0
tensorboard==1.9.0
tensorboardX==2.2
tensorflow-gpu==1.9.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.8.25
toml==0.10.2
toolz==0.11.1
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==6.1
tqdm==4.15.0
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
update==0.0.1
urllib3==1.25.10
visdom==0.1.8.3
wcwidth==0.2.5
websocket-client==1.0.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
zope.interface==5.1.0
